Исправление ошибок + Docker для развертывания

+++ TactiX/Models/ViewModels/TrainingStageCreateDto.cs
@@ -20,25 +20,25 @@ namespace TactiX.Models.ViewModels
         public double HitFactor { get; set; }
 
         [Required]
-        public long NumOfSpots { get; set; }
+        public int NumOfSpots { get; set; }
 
         [Required]
-        public long NumOfPoppers { get; set; }
+        public int NumOfPoppers { get; set; }
 
         [Required]
-        public long NumOfPlates { get; set; }
+        public int NumOfPlates { get; set; }
 
         [Required]
-        public long AlphasCount { get; set; }
+        public int AlphasCount { get; set; }
 
         [Required]
-        public long DeltasCount { get; set; }
+        public int DeltasCount { get; set; }
 
         [Required]
-        public long CharliesCount { get; set; }
+        public int CharliesCount { get; set; }
 
         [Required]
-        public long MissesCount { get; set; }
+        public int MissesCount { get; set; }
 
         public string StageTimeInput { get; set; }
